Change Location properties from int to float in ArtboardElement classes

+++ ITBees.DragArtboard/Interfaces/Models/ArtboardElementIm.cs
@@ -5,9 +5,9 @@ public class ArtboardElementIm
     public Guid ArtboardGuid { get; set; }
     public Guid ArtboardToolboxElementGuid { get; set; }
     public string CustomSerializedObject { get; set; }
-    public int LocationX { get; set; }
-    public int LocationY { get; set; }
-    public int LocationZ { get; set; }
+    public float LocationX { get; set; }
+    public float LocationY { get; set; }
+    public float LocationZ { get; set; }
     public int ZIndex { get; set; }
     public int Width { get; set; }
     public int Height { get; set; }
